requirements.txt
beautifulsoup4==4.6.0
certifi==2018.1.18
chardet==3.0.4
Django==2.0.4
EasyProcess==0.2.3
future==0.16.0
Google-Search-API==1.1.14
idna==2.6
image==1.5.20
modules==1.0.0
mss==3.2.0
multidict==4.1.0
nltk==3.2.5
Pillow==5.1.0
pyscreenshot==0.4.2
pytz==2018.3
PyYAML==3.12
requests==2.18.4
selenium==2.53.6
six==1.11.0
tesserocr==2.2.2
Unidecode==1.0.22
urllib3==1.22
vcrpy==1.11.1
wrapt==1.10.11
xlib==0.21
yarl==1.1.1
